Add crawl_delay
Saw this in robots.txt of the sites I want to crawl. One of them specifies crawl-delay globally, which urllib.robotparser doesn't support, and neither should block crawling on specific URLs.

So, at this stage this covers everything needed without the overhead of a full robots.txt parser.

If parallelizing the workers, then a more sophisticated throttler implementation will be needed, as the crawl delay should be global.

I'm also assuming that the crawl is going to stay within a single site, rather than branch out which would require another robots.txt check. That's true for my initial use case, but not necessarily so if someone wrote their own handler for a different use case.