Add handler for index pages in The European Cultivated Potato Database
Nothing here to scrape, just want to

Adding terms.html as evidence that this data is available to the public, as long as it's acknowledged.

Think I might want to move the resolution of relative URLs into crawler, as it's going to happen a lot and it doesn't make sense for each handler to have to do the same thing.