Extract links to europotato database
Have a feeling this will come in handy for matching records.

I don't think it's 100% reliable, but it's a start. The links don't always work, which makes me think this is a name from the pedigree database that they're hoping will be in the europotato one.

Didn't do this the other way because europotato is explicitly linking to the pedigree's search page using the europotato name, so it doesn't guarantee any match.