Restrict to single in progress URL
The new implementation is less disk intensive but more linear. It assumes we have a single element being processed at a time. Otherwise, if we have two in progress URLs, and mark one as complete/failed, we'll increment the queue counter. But on restore, that might have mean skipping over the other incomplete URL. Not what we want. This makes the check have a tighter bound.

Makes me think this should all be sqlite honestly.